+++ src/components/controllers/Projects.vue
@@ -109,40 +109,46 @@ export default {
     };
   },
   created() {
-    
-    ApiService.getProjects().then((result) => {
-      this.projectsList = result;
-    })
+    if (process.env.VUE_APP_FAKE_PROJECT_CALLS !== "true") {
+      ApiService.getProjects().then((result) => {
+        this.projectsList = result;
+      })
+    } else {
+      this.projectsList = [
+        {"project_id":"23","name":"Filmy Cue 1","description":"dark, brooding","song_count":"7"},
+        {"project_id":"31","name":"foo proj","description":"desc","song_count":"0"},
+        {"project_id":"30","name":"Reboot","description":"Next Level","song_count":"0"}
+      ];
+    }
 
-    //this.projectsList = [
-    //  {"project_id":"23","name":"Filmy Cue 1","description":"dark, brooding","song_count":"7"},
-    //  {"project_id":"31","name":"foo proj","description":"desc","song_count":"0"},
-    //  {"project_id":"30","name":"Reboot","description":"Next Level","song_count":"0"}
-    //];
+
+
   },
   methods: {
     setSelectedProject(project) {
 
-      ApiService.getProjectDetails(project.project_id).then((data) => {
-        this.selectedProject = data;
+      if (process.env.VUE_APP_FAKE_PROJECT_CALLS !== "true") {
+        ApiService.getProjectDetails(project.project_id).then((result) => {
+        console.debug("project data: ", result)
+        this.selectedProject = JSON.parse(JSON.stringify(result));
       })
-      /* 
-      this.selectedProject = {
-        "project_id":"23",
-        "customer_id":"14",
-        "name":"Filmy Cue 1",
-        "description":"dark, brooding",
-        "songs":[ 
-          {"song_id":"126", "snippet_file": "3OR4TIMESSNIP.mp3", "part_number":"50517001","title":"3 Or 4 Times_Full","mix_id":"12","remix_capable":"1","duration":"00:03:59","tempo":"Legato","description":"Solo Piano concerto, elegant, emotional","tempo_type_id":"3","song_key":"N/A","meter":"4/4","copyright_number":"Phillip T. Moody","is_public_domain":"0","time_period_id":"1","isrc_no":null,"pro_song_reg_no":null,"is_exclusive":"1","project_song_id":"267","project_id":"23","comments":"cool"},
-          {"song_id":"3296", "snippet_file": "3OR4TIMESSNIP.mp3", "part_number":"60615001","title":"A Nest In The Family Tree_Full","mix_id":"1","remix_capable":"1","duration":"00:00:59","tempo":"95","description":"Poignant downtempo orchestral instrumental, sentimental and bittersweet","tempo_type_id":"3","song_key":"DM","meter":"4/4","copyright_number":"Chris Fuller","is_public_domain":"0","time_period_id":"12","isrc_no":null,"pro_song_reg_no":null,"is_exclusive":null,"project_song_id":"177","project_id":"23","comments":"This one is dark!"},
-          {"song_id":"3560","snippet_file": "3OR4TIMESSNIP.mp3", "part_number":"61201001","title":"Aarons Fantastic Fandango_Full","mix_id":"1","remix_capable":"0","duration":"00:03:08","tempo":"100","description":"Easy acoustic guitar strumming and congas, mellow feel, piano adds flavor","tempo_type_id":"2","song_key":"DM","meter":"4/4","copyright_number":"David Graham","is_public_domain":"0","time_period_id":"1","isrc_no":null,"pro_song_reg_no":null,"is_exclusive":null,"project_song_id":"272","project_id":"23","comments":""},
-          {"song_id":"20","snippet_file": "3OR4TIMESSNIP.mp3","part_number":"50108001","title":"Be A Lover_Full","mix_id":"1","remix_capable":"1","duration":"00:03:06","tempo":"95","description":"Male lead vocal, soft rock in the style of Bruce Hornsby","tempo_type_id":"2","song_key":"Dm","meter":"4/4","copyright_number":"PAu2-015-180","is_public_domain":"0","time_period_id":"1","isrc_no":null,"pro_song_reg_no":null,"is_exclusive":"1","project_song_id":"269","project_id":"23","comments":"Good for the middle."},
-          {"song_id":"6516","snippet_file": "3OR4TIMESSNIP.mp3", "part_number":"67407002","title":"MSO Fixing It_Alt","mix_id":"4","remix_capable":"0","duration":"00:02:52","tempo":"98","description":"Modern pop rock which sounds like Coldplay's 'Fix You'. Also sounds like Adele, U2, Keane, Radiohead & Nickelback. Long version.","tempo_type_id":"3","song_key":"EM","meter":"4/4","copyright_number":"Dean James Wagg","is_public_domain":"0","time_period_id":"1","isrc_no":"","pro_song_reg_no":"","is_exclusive":"0","project_song_id":"271","project_id":"23","comments":""},
-          {"song_id":"6517","snippet_file": "3OR4TIMESSNIP.mp3", "part_number":"67408001","title":"MSO Fly To The Stars_Full","mix_id":"1","remix_capable":"0","duration":"00:02:15","tempo":"110","description":"Beautiful delicate & calm composition for Piano & strings.","tempo_type_id":"2","song_key":"CM","meter":"4/4","copyright_number":"Dean James Wagg","is_public_domain":"0","time_period_id":"1","isrc_no":"","pro_song_reg_no":"","is_exclusive":"0","project_song_id":"270","project_id":"23","comments":"flying is good"},
-          {"song_id":"6","snippet_file": "3OR4TIMESSNIP.mp3", "part_number":"50103001","title":"Now I Lay Me Down_Full","mix_id":"1","remix_capable":"1","duration":"00:03:17","tempo":"80","description":"Female lead vocal, a prayer, hip beat with guitar & synth pad","tempo_type_id":"2","song_key":"AbM","meter":"4/4","copyright_number":"SRu559-370","is_public_domain":"0","time_period_id":"1","isrc_no":null,"pro_song_reg_no":null,"is_exclusive":"1","project_song_id":"268","project_id":"23","comments":"Good for the first song."}]
-        }
-      */
-
+      
+      } else {
+        this.selectedProject = {
+          "project_id":"23",
+          "customer_id":"14",
+          "name":"Filmy Cue 1",
+          "description":"dark, brooding",
+          "songs":[ 
+            {"song_id":"126", "snippet_file": "3OR4TIMESSNIP.mp3", "part_number":"50517001","title":"3 Or 4 Times_Full","mix_id":"12","remix_capable":"1","duration":"00:03:59","tempo":"Legato","description":"Solo Piano concerto, elegant, emotional","tempo_type_id":"3","song_key":"N/A","meter":"4/4","copyright_number":"Phillip T. Moody","is_public_domain":"0","time_period_id":"1","isrc_no":null,"pro_song_reg_no":null,"is_exclusive":"1","project_song_id":"267","project_id":"23","comments":"cool"},
+            {"song_id":"3296", "snippet_file": "3OR4TIMESSNIP.mp3", "part_number":"60615001","title":"A Nest In The Family Tree_Full","mix_id":"1","remix_capable":"1","duration":"00:00:59","tempo":"95","description":"Poignant downtempo orchestral instrumental, sentimental and bittersweet","tempo_type_id":"3","song_key":"DM","meter":"4/4","copyright_number":"Chris Fuller","is_public_domain":"0","time_period_id":"12","isrc_no":null,"pro_song_reg_no":null,"is_exclusive":null,"project_song_id":"177","project_id":"23","comments":"This one is dark!"},
+            {"song_id":"3560","snippet_file": "3OR4TIMESSNIP.mp3", "part_number":"61201001","title":"Aarons Fantastic Fandango_Full","mix_id":"1","remix_capable":"0","duration":"00:03:08","tempo":"100","description":"Easy acoustic guitar strumming and congas, mellow feel, piano adds flavor","tempo_type_id":"2","song_key":"DM","meter":"4/4","copyright_number":"David Graham","is_public_domain":"0","time_period_id":"1","isrc_no":null,"pro_song_reg_no":null,"is_exclusive":null,"project_song_id":"272","project_id":"23","comments":""},
+            {"song_id":"20","snippet_file": "3OR4TIMESSNIP.mp3","part_number":"50108001","title":"Be A Lover_Full","mix_id":"1","remix_capable":"1","duration":"00:03:06","tempo":"95","description":"Male lead vocal, soft rock in the style of Bruce Hornsby","tempo_type_id":"2","song_key":"Dm","meter":"4/4","copyright_number":"PAu2-015-180","is_public_domain":"0","time_period_id":"1","isrc_no":null,"pro_song_reg_no":null,"is_exclusive":"1","project_song_id":"269","project_id":"23","comments":"Good for the middle."},
+            {"song_id":"6516","snippet_file": "3OR4TIMESSNIP.mp3", "part_number":"67407002","title":"MSO Fixing It_Alt","mix_id":"4","remix_capable":"0","duration":"00:02:52","tempo":"98","description":"Modern pop rock which sounds like Coldplay's 'Fix You'. Also sounds like Adele, U2, Keane, Radiohead & Nickelback. Long version.","tempo_type_id":"3","song_key":"EM","meter":"4/4","copyright_number":"Dean James Wagg","is_public_domain":"0","time_period_id":"1","isrc_no":"","pro_song_reg_no":"","is_exclusive":"0","project_song_id":"271","project_id":"23","comments":""},
+            {"song_id":"6517","snippet_file": "3OR4TIMESSNIP.mp3", "part_number":"67408001","title":"MSO Fly To The Stars_Full","mix_id":"1","remix_capable":"0","duration":"00:02:15","tempo":"110","description":"Beautiful delicate & calm composition for Piano & strings.","tempo_type_id":"2","song_key":"CM","meter":"4/4","copyright_number":"Dean James Wagg","is_public_domain":"0","time_period_id":"1","isrc_no":"","pro_song_reg_no":"","is_exclusive":"0","project_song_id":"270","project_id":"23","comments":"flying is good"},
+            {"song_id":"6","snippet_file": "3OR4TIMESSNIP.mp3", "part_number":"50103001","title":"Now I Lay Me Down_Full","mix_id":"1","remix_capable":"1","duration":"00:03:17","tempo":"80","description":"Female lead vocal, a prayer, hip beat with guitar & synth pad","tempo_type_id":"2","song_key":"AbM","meter":"4/4","copyright_number":"SRu559-370","is_public_domain":"0","time_period_id":"1","isrc_no":null,"pro_song_reg_no":null,"is_exclusive":"1","project_song_id":"268","project_id":"23","comments":"Good for the first song."}]
+          }
+      }
     },
     deleteProject(project) {
       
@@ -169,11 +175,14 @@ export default {
               return;
             }
 
-            ApiService.deleteProject(project.project_id).then(() => {
-              this.projectsList = this.projectsList.filter(p => p.project_id !== project.project_id);
-            })
+            if (process.env.VUE_APP_FAKE_PROJECT_CALLS !== "true") {
+              ApiService.deleteProject(project.project_id).then(() => {
+                this.projectsList = this.projectsList.filter(p => p.project_id !== project.project_id);
+              })
 
-            //this.projectsList = this.projectsList.filter(p => p.project_id !== project.project_id);
+            } else {
+              this.projectsList = this.projectsList.filter(p => p.project_id !== project.project_id);
+            }
           })
           .catch(() => {
             // An error occurred
